Persist logs to files in the host
This persist the audit logs to the host like one would usually expect
from auditd.

To achieve this, we needed to change the auditd command to use the -f
flag. This stops the process from forking, while keeping the log
behavior.

auditd also requires /dev/log to work, so we mount it from the host;
this allows us to also output the status of auditd to the host's logs
(it ends up in the journal).